+++ src/main/java/com/medify/service/UserServiceImpl.java
@@ -4,10 +4,14 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import com.medify.entity.User;
 import com.medify.repository.UserRepository;
+
+import javax.transaction.Transactional;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -26,7 +30,7 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public User getUserById(Integer id) {
+	public User getUserById(Long id) {
 		Optional<User> user=userRepository.findById(id);
 		if(user.isPresent())
 		{
@@ -36,13 +40,14 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public User updateUserById(Integer id, User user) {
+	public User updateUserById(Long id, User user) {
 		return userRepository.save(user);
 	}
 
 	@Override
-	public void deleteUserById(Integer id) {
+	public void deleteUserById(Long id) {
 		userRepository.deleteById(id);
 	}
+	
 
 }
